+++ tests/packerlicious/test_post_processor_vagrant.py
@@ -1,0 +1,9 @@
+import packerlicious.post_processor as post_processor
+
+
+class TestVagrantPostProcessor(object):
+
+    def test_no_required_fields(self):
+        b = post_processor.Vagrant()
+
+        b.to_dict()