+++ CHANGELOG.md
@@ -12,4 +12,5 @@
 * post_processor/GoogleComputeExport: add support for packer's google compute image exporter post-processor
 * post_processor/Manifest: add support for packer's manifest post-processor
 * post_processor/ShellLocal: add support for packer's shell local post-processor
+* post_processor/Vagrant: add support for packer's vagrant post-processor
 * provisioner/file: add support for packer's file provisioner